Do not set request-backend to url-retrieve
Don't force using `url-retrieve` as `request-backend`. The user might
want to use `curl` as `request-backend`.

At the time of this commit there is an Emacs bug with `url-retrieve` and
non-ascii characters in the request:
http://debbugs.gnu.org/cgi/bugreport.cgi?bug=24117#23

To work around the bug the `request-backend` can be set to `curl`.

When sending the request synchronously while using `curl`, Emacs prints
a Warning (`Blocking call to accept-process-output with quit
inhibited!!`). To suppress the warning we need to surround the
blocking `request` call with `with-local-quit` to make quitting the
process with `C-g` work.